apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
PyYAML==6.0.2
langdetect==1.0.9
openai==1.51.2
